Lesson_5

+++ RestarauntBooking/Restaurant.cs
@@ -12,7 +12,7 @@ namespace RestarauntBooking
 
         public Restaurant()
         {
-            for (int i = 1; i <= 10; i++)
+            for (ushort i = 1; i <= 10; i++)
             {
                 _tables.Add(new Table(i));
             }
